upload tach check diag info
