+++ python/tach/cli.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import argparse
-import re
 import sys
 from dataclasses import dataclass, field
 from enum import Enum
@@ -28,6 +27,7 @@ from tach.filesystem import install_pre_commit
 from tach.logging import LogDataModel, logger
 from tach.modularity import export_modularity
 from tach.parsing import parse_project_config
+from tach.parsing.config import extend_and_validate
 from tach.report import external_dependency_report, report
 from tach.show import (
     generate_module_graph_dot_file,
@@ -491,29 +491,6 @@ def check_cache_for_action(
             exit_code=cache_result[1],
         )
     return CachedOutput(key=cache_key)
-
-
-def extend_and_validate(
-    exclude_paths: list[str] | None,
-    project_excludes: list[str],
-    use_regex_matching: bool,
-) -> list[str]:
-    if exclude_paths is not None:
-        exclude_paths.extend(project_excludes)
-    else:
-        exclude_paths = project_excludes
-
-    if not use_regex_matching:
-        return exclude_paths
-
-    for exclude_path in exclude_paths:
-        try:
-            re.compile(exclude_path)
-        except re.error:
-            raise ValueError(
-                f"Invalid regex pattern: {exclude_path}. If you meant to use glob matching, set 'use_regex_matching' to false in your .toml file."
-            )
-    return exclude_paths
 
 
 def tach_check(
