+++ python/tach/extension.pyi
@@ -61,7 +61,7 @@ def check(
     project_root: Path,
     project_config: ProjectConfig,
     exclude_paths: list[str],
-) -> CheckResult: ...
+) -> CheckDiagnostics: ...
 def sync_dependency_constraints(
     project_root: Path,
     project_config: ProjectConfig,
@@ -92,7 +92,7 @@ class BoundaryError:
     import_mod_path: str
     error_info: ErrorInfo
 
-class CheckResult:
+class CheckDiagnostics:
     errors: list[BoundaryError]
     deprecated_warnings: list[BoundaryError]
     warnings: list[str]
